Relocate Groovy classes from `org.grails.cli.groovy` to `org.grails.cli.compiler.groovy`
