+++ grace-shell/src/main/groovy/org/grails/cli/compiler/DependencyManagementBomTransformation.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2023 the original author or authors.
+ * Copyright 2012-2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -51,7 +51,7 @@ import org.springframework.core.annotation.Order;
 
 import org.grails.cli.compiler.dependencies.MavenModelDependencyManagement;
 import org.grails.cli.compiler.grape.DependencyResolutionContext;
-import org.grails.cli.groovy.DependencyManagementBom;
+import org.grails.cli.compiler.groovy.DependencyManagementBom;
 
 /**
  * {@link ASTTransformation} for processing
