+++ grace-shell/src/main/groovy/org/grails/cli/compiler/GenericBomAstTransformation.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2023 the original author or authors.
+ * Copyright 2012-2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -35,7 +35,7 @@ import org.codehaus.groovy.control.SourceUnit;
 import org.codehaus.groovy.transform.GroovyASTTransformation;
 import org.springframework.core.Ordered;
 
-import org.grails.cli.groovy.DependencyManagementBom;
+import org.grails.cli.compiler.groovy.DependencyManagementBom;
 
 /**
  * A base class that lets plugin authors easily add additional BOMs to all apps. All the
