+++ grace-shell/src/main/groovy/org/grails/cli/compiler/groovy/GroovyTemplate.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2019 the original author or authors.
+ * Copyright 2012-2026 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,8 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package org.grails.cli.groovy;
+package org.grails.cli.compiler.groovy;
 
 import java.io.File;
 import java.io.IOException;
